// packages/backend/src/dtos/contract.dto.ts
export interface IContract {
    nrInst: string
    nrAgencia: string
    cdClient: string
    nmClient: string
    nrCpfCnpj: string
    nrContrato: string
    dtContrato: Date
    qtPrestacoes: string
    vlTotal: string
    cdProduto: string
    dsProduto: string
    cdCarteira: string
    dsCarteira: string
    nrProposta: string
    nrPresta: string
    tpPresta: string
    nrSeqPre: string
    dtVctPre: Date
    vlPresta: string
    vlMora: string
    vlMulta: string
    vlOutAcr: string
    vlIof: string
    vlDescon: string
    vlAtual: string
    idSituac: string
    idSitVen: string
    vlMovimento: boolean
    nrTypeDocument: string
    vlInstallmentValue: boolean
  }